payment system for both flight and hotel bookings
