+++ Backend/package.json
@@ -32,7 +32,9 @@
     "nodemailer": "^6.9.13",
     "passport": "^0.7.0",
     "passport-google-oauth20": "^2.0.0",
-    "redis": "^4.6.13"
+    "redis": "^4.6.13",
+    "stripe": "^14.23.0",
+    "uuid4": "^2.0.3"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.0",
